Initial commit of the EosOverclockHelper.
This application simply runs on boot and checks for the result of the
overclocking script that is run on init. If it finds that the script failed,
then it creates a notification saying so. When this notification is pressed,
the user is taken to the Eos Overclock Settings page, where they can set
their values again.